docs: Spliting the project on several folders
